Moved getParamsFromText() from parsexmlutils.cpp to mqlutils.cpp
