+++ common/parsexmlutils.cpp
@@ -1855,32 +1855,3 @@ bool parseReportParameter(const QDomElement & elemSource, ORReportData & reportT
 
   return true;
 }
-
-// TODO: It would be better to alter the metasql parser to do this for us
-QStringList getParamsFromText(const QString p)
-{
-  QStringList result;
-
-  QRegExp tagre("(<\\?[^?]*\\?>)");
-  for (int i = 0; (i = tagre.indexIn(p, i)) != -1; i += tagre.matchedLength())
-  {
-    QString tag = tagre.cap(1);
-    if (DEBUG)
-      qDebug("getParamsFromMetaSQLText looking at %s", qPrintable(tag));
-
-    QRegExp paramre("(['\\\"])([^'\\\"]+)\\1");
-    for (int j = 0; (j = paramre.indexIn(tag, j)) != -1; j += paramre.matchedLength())
-    {
-      QString param = paramre.cap(2);   // the \w+, not the ['"]
-      if (DEBUG)
-        qDebug("getParamsFromMetaSQLText found %s", qPrintable(param));
-      result.append(param);
-    }
-  }
-
-  result.sort();
-  result.removeDuplicates();
-
-  return result;
-}
-
